moved filter to hooks callbacks
